backend/requirements.txt
numpy
tensorflow
flask
flask-cors
Pillow
